\name{DifferentialRegulation-package}
\alias{DifferentialRegulation-package}
\alias{DifferentialRegulation}
\docType{package}
\title{
\packageTitle{DifferentialRegulation}
}
\description{
\packageDescription{DifferentialRegulation}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DifferentialRegulation}
Questions relative to DifferentialRegulation should be reported as a new issue at \href{https://github.com/SimoneTiberi/DifferentialRegulation/issues}{https://github.com/SimoneTiberi/DifferentialRegulatio/issues}.

To access the vignettes, type: \code{browseVignettes("DifferentialRegulation")}.

\packageIndices{DifferentialRegulation}
}
\author{
\packageAuthor{DifferentialRegulation}
Maintainer: \packageMaintainer{DifferentialRegulation}
}
\keyword{ package }
