% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialRegulation.R
\name{DifferentialRegulation}
\alias{DifferentialRegulation}
\title{Discover differentially regulated genes}
\usage{
DifferentialRegulation(
  sce,
  EC_list = NULL,
  samples_design,
  sample_col_name = "sample",
  group_col_name = "group",
  sce_cluster_name = "cell_type",
  min_cells_per_cluster = 100,
  min_counts_per_gene_per_group = 20,
  n_cores = NULL,
  N_MCMC = 2000,
  burn_in = 500,
  min_counts_ECs = 0
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment} object, computed via \code{load_USA}.}

\item{EC_list}{a \code{list}, computed via \code{load_EC}.}

\item{samples_design}{TODO}

\item{sample_col_name}{TODO}

\item{group_col_name}{TODO}

\item{sce_cluster_name}{TODO}

\item{min_cells_per_cluster}{TODO}

\item{min_counts_per_gene_per_group}{TODO}

\item{n_cores}{TODO}

\item{N_MCMC}{TODO}

\item{burn_in}{TODO}

\item{min_counts_ECs}{TODO}
}
\value{
A \code{data.frame} object.
}
\description{
\code{DifferentialRegulation} identified differentially regulated genes between two conditions 
(e.g., healthy vs. disease or treated vs. untreated) in each cluster of cells.
Parameters are inferred via Markov chain Monte Carlo (MCMC) techniques and a differential testing is performed 
via a multivariate Wald test on the posterior densities of the group-level USA (Unspliced, Spliced and Ambiguous) counts relative abundance.
}
\examples{
TODO

}
\seealso{
\code{\link{load_EC}}, \code{\link{load_USA}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
