% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_EC.R
\name{load_EC}
\alias{load_EC}
\title{Create a list containing the equivalence classes objects object}
\usage{
load_EC(
  path_to_EC_counts,
  path_to_EC,
  path_to_cell_id,
  path_to_gene_id,
  sample_ids
)
}
\arguments{
\item{path_to_EC_counts}{a vector of length equals to the number of samples: 
each element indicates the path to the equivalence classes counts of the respective sample (i.e., geqc_counts.mtx file).}

\item{path_to_EC}{a vector of length equals to the number of samples: 
each element indicates the path to the equivalence classes of the respective sample (i.e., gene_eqclass.txt.gz file).}

\item{path_to_cell_id}{a vector of length equals to the number of samples: 
each element indicates the path to the cell ids of the respective sample (i.e., quants_mat_rows.txt file).}

\item{path_to_gene_id}{a vector of length equals to the number of samples: 
each element indicates the path to the gene ids of the respective sample (i.e., quants_mat_cols.txt file).}

\item{sample_ids}{a vector of length equals to the number of samples: 
each element indicates the name of the sample.}
}
\value{
A \code{list} object.
}
\description{
\code{load_EC} imports the equivalence classes (computed by alevin-fry), and stores them into a list.
}
\examples{
TODO

}
\seealso{
\code{\link{load_USA}}, \code{\link{DifferentialRegulation}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
