% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_USA.R
\name{load_USA}
\alias{load_USA}
\title{Create a list containing the equivalence classes objects object}
\usage{
load_USA(path_to_counts, path_to_cell_id, path_to_gene_id, sample_ids)
}
\arguments{
\item{path_to_counts}{a vector of length equals to the number of samples: 
each element indicates the path to the USA estimated count matrix of the respective sample (i.e., quants_mat.mtx file).}

\item{path_to_cell_id}{a vector of length equals to the number of samples: 
each element indicates the path to the cell ids of the respective sample (i.e., quants_mat_rows.txt file).}

\item{path_to_gene_id}{a vector of length equals to the number of samples: 
each element indicates the path to the gene ids of the respective sample (i.e., quants_mat_cols.txt file).}

\item{sample_ids}{a vector of length equals to the number of samples: 
each element indicates the name of the sample.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
\code{load_USA} imports the estimated USA (Unspliced, Spliced and Ambiguous) counts (computed by alevin-fry), 
and stores them into a \code{SingleCellExperiment} object.
}
\examples{
TODO

}
\seealso{
\code{\link{load_EC}}, \code{\link{DifferentialRegulation}}
}
\author{
Simone Tiberi \email{simone.tiberi@uzh.ch}
}
